* Hokay, so. I've finally implemented a state machine, driven by events. If I want to change a screen I throw a ChangeState event with the appropriate enum to access the new state.
* I have now implemented a super primitive main menu. Options menu doesn't work quite yet, everything is WIP, everything looks ass, but it's functional enough.

* Done quite a bit of restructuring both in code and the vsproj filters, but that's to be expected, since the game is now  driven by a state machine instead of just.... running in the main gameloop without any logic for changing any screens or anything.

* Added bunch of functions for "resetting" gameplay screen, so that you get a fresh world every time you exit to main menu and enter again.

+++ Engine/frustumclass.cpp
@@ -58,7 +58,7 @@ void FrustumClass::SetInternals( float aspectRatio, float FOV, float nearZ, floa
 	Wfar = Hfar * ratio 
 	*/
 
-	float tanVal = 2.0f * tan((ANG2RAD * FOV) / 2);
+	float tanVal = 2.0f * (float)tan((ANG2RAD * FOV) / 2);
 
 	// compute width and height of the near and far plane sections
 	farHeight = tanVal * farZ;
